Use short for numQues
Number of question would not exceed 100 for a session (generally).
Byte would be more appropriate primitive for numQues; But for the sake of
expansion later on, it would be reasonable to use short.

Also this differentiation helps in setting constructors later on. If numQues and
maxRange would both be int, setting them in the constructor would be difficult.

+++ src/test/java/org/arithmetic/SessionTest.java
@@ -37,25 +37,25 @@ public class SessionTest {
 
     @Test(expected = AppException.class)
     public void zeroNumQuesTest() throws AppException {
-        int n = 0;
+        short n = 0;
         Session s = new Session();
         s.setNumQues(n);
     }
 
     @Test(expected = AppException.class)
     public void negNumQuesTest() throws AppException {
-        int n = -5;
+        short n = -5;
         Session s = new Session();
         s.setNumQues(n);
     }
 
     @Test
     public void posNumQuesTest() throws AppException {
-        int n = 5;
+        short n = 5;
         Session s = new Session();
         s.setNumQues(n);
 
-        int testRes = s.getNumQues();
+        short testRes = s.getNumQues();
 
         assertEquals(testRes, n);
     }
